Working refactor for 4 channels

+++ include/globals.h
@@ -3,7 +3,7 @@
  * 
  * Project globals, mostly pre-processor stuff
  */
-#define NUM_CHANNELS 4
-#define SKIP_SIZE 2
-#define V_MAX 1023
-#define HYSTERESIS 3
+#define NUM_CHANNELS (uint8_t)4
+#define SKIP_SIZE (uint8_t)2
+#define V_MAX (uint16_t)1023
+#define HYSTERESIS (uint8_t)3
